Added Memory feature

+++ scheduler.c
@@ -2,26 +2,31 @@
 #include "headers.h"
 #include "Defined_DS.h"
 
-// -------------------Closing the scheduler---------------------
+// ---------------- Closing the scheduler --------------------
 void closeScheduler(int sig_num);
 
-// -------------------  Output file functions-------------------
+// ----------------  Output file functions -------------------
 void addToLog(struct PCB pbcblock,int time);
 void addToPerf(struct PCB* pcb, int process_count);
+void MemLog(struct MemoryBlock mem, int time);
 
-// -------------------  HPF functions---------------------------
+// -----------------  HPF functions --------------------------
 void HPF(int process_count);
 void HPFhandler(int sig_num);
 
-// -------------------  RR functions---------------------------
+// -----------------  RR functions ---------------------------
 void RR(int process_count, int quantum);
 
-// -------------------  SRTN functions---------------------------
+// ----------------  SRTN functions --------------------------
 void SRTN(int process_count);
-//----------------------------------------------------------------
+
+//----------------- Memory Functions -------------------------
+MemBlock* allocateMem(MemBlock* root, int size, int process_id);
+bool freeMem(MemBlock* root, int process_id);
 
 FILE *logFile;
 FILE *perfFile;
+FILE *MemFile;
 
 key_t msgkey; 
 int attach;
@@ -29,11 +34,14 @@ int attach;
 struct PCB* pcb;
 struct PCB* RR_pcb;
 
+MemBlock* mem;
+
 double util = 0;
 int finish;
 
 int main(int argc, char *argv[]) //1- alognumber,2- process_count,3- quantum
 {
+    
     initClk();
     signal(SIGINT,closeScheduler);
 
@@ -47,25 +55,27 @@ int main(int argc, char *argv[]) //1- alognumber,2- process_count,3- quantum
 
     logFile = fopen("scheduler.log", "w");
     perfFile = fopen("scheduler.perf", "w");
-    
+    MemFile = fopen("memory.log", "w");
 
-    if (logFile == NULL || perfFile == NULL) {
+
+    if (logFile == NULL || perfFile == NULL || MemFile == NULL) {
         perror("Error opening output files");
         msgctl(attach,IPC_RMID,NULL);
         exit(1);
     }
     
     fprintf(logFile,"#At time x process y state arr w total z remain y wait k\n");
+    fprintf(MemFile,"#At time x allocated y bytes for process z from i to j\n");
 
     int algo = atoi(argv[1]);
     int processes_count = atoi(argv[2]);
 
+    mem = createMemBlock(0,1024,10);
 
     switch (algo)
     {
         case 1: 
             RR(processes_count,atoi(argv[3]));
-            finish = getClk();
             addToPerf(RR_pcb,processes_count);
             break;
         case 2:
@@ -89,26 +99,40 @@ int completed;
 struct Process runningProcess = {0,0,0,0};
 int pcbidx = 0;     //tracks the current running process index
 int pausetime = 0;
-
+PriorityQueue *BlocklistPri;
 void HPF(int process_count)
 {
-      
+
+    completed = 0; 
     signal(SIGCHLD,HPFhandler);
     //tracks the number of processes that own a PCB  
     int pcbSize = 0;
     Process_queue = createPriorityQueue(process_count);
     pcb = (struct PCB*)malloc(sizeof(PCB) * process_count);
     int pid;
+    BlocklistPri = createPriorityQueue(process_count);
 
     while(completed < process_count)
     {
-        while(msgrcv(attach,&newMessage,sizeof(newMessage.p),0,IPC_NOWAIT) != -1)
+        int time = getClk();
+        while(msgrcv(attach,&newMessage,sizeof(newMessage.p),0,IPC_NOWAIT) != -1 && time == getClk())
         {
             printf("Process %d received at time %d\n",newMessage.p.id,getClk());
-            insertPriorityPriorityQueue(Process_queue, newMessage.p);
             pcb[pcbSize].processstate = ready;                  //The test generator sets process ids from 1->process_count in order                                         
             pcb[pcbSize].remainingtime = newMessage.p.runtime;  //We will access PCB of a certain process through its id
-            pcb[pcbSize++].p = newMessage.p;    
+            pcb[pcbSize].p = newMessage.p;  
+            
+            MemBlock* allocate;
+            allocate = allocateMem(mem,newMessage.p.memsize,newMessage.p.id);
+            if (allocate != NULL)
+            {
+                insertPriorityPriorityQueue(Process_queue, newMessage.p);
+                pcb[pcbSize].memblock = allocate;
+            }
+            else
+                insertPriorityPriorityQueue(BlocklistPri,newMessage.p);
+
+            pcbSize++;
         }
 
         if (Process_queue->size > 0)
@@ -149,31 +173,62 @@ void HPFhandler(int sig_num)
     pcb[pcbidx].TA = pcb[pcbidx].finishedtime - runningProcess.arrival_time ;
     pcb[pcbidx].WTA = pcb[pcbidx].TA/runningProcess.runtime;
     completed++;
-    addToLog(pcb[pcbidx],getClk());    
+    addToLog(pcb[pcbidx],getClk());  
+    bool isfree = freeMem(mem,runningProcess.id);
+    if(!isfree)
+    {
+        printf("Error freeing memory for process %d\n",runningProcess.id);
+    }
+
+    //Tries to allocate memory for the Blocklist processes
+    MemBlock* allocate = NULL;
+
+    while (BlocklistPri->size > 0)
+    {
+        allocate = allocateMem(mem,BlocklistPri->array->memsize,BlocklistPri->array->id);
+        if (allocate != NULL)
+        {
+            Process p = removePriorityPriorityQueue(BlocklistPri);
+            insertPriorityPriorityQueue(Process_queue, p);
+            pcb[p.id - 1].memblock = allocate;
+        }
+        else
+            break;
+    }
 }
 
 
 //////////////////////////////////// Round Robin ///////////////////////////////////////////////////////
 CircularQueue* RR_queue;
 int RR_completed = 0;
-struct Process RR_runningProcess = {0,0,0,0};
+struct Process RR_runningProcess = {0,0,0,0,0};
 bool Run = false;
 int RR_startTime = 0;
-int RR_current_pid = -1;
-
+CircularQueue *BlockList;
 void RR(int process_count, int quantum) {   
     RR_queue = createCircularQueue(process_count);
     RR_pcb = (struct PCB*)malloc(sizeof(PCB) * process_count);
     memset(RR_pcb, 0, sizeof(PCB) * process_count);
+    BlockList = createCircularQueue(process_count);
 
     while (RR_completed < process_count) {
 
-        while (msgrcv(attach, &newMessage, sizeof(newMessage.p), 0, IPC_NOWAIT) != -1) {
+        while (msgrcv(attach, &newMessage, sizeof(newMessage.p), 0, IPC_NOWAIT) != -1)
+        {
             int idx = newMessage.p.id - 1;       
             RR_pcb[idx].p = newMessage.p;
             RR_pcb[idx].remainingtime = newMessage.p.runtime;
             RR_pcb[idx].processstate = ready;
-            enqueueCircularQueue(RR_queue, newMessage.p);
+
+            MemBlock* allocate;
+            allocate = allocateMem(mem,newMessage.p.memsize,newMessage.p.id);
+            if(allocate != NULL)
+            {
+                enqueueCircularQueue(RR_queue, newMessage.p);
+                RR_pcb[idx].memblock = allocate;
+            }
+            else
+                enqueueCircularQueue(BlockList,newMessage.p);
         }
 
         int current_time = getClk();
@@ -181,7 +236,6 @@ void RR(int process_count, int quantum) {
         if(Run && (current_time - RR_startTime >= quantum)) {
             if (RR_pcb[pcbidx].remainingtime > quantum)
                 kill(RR_pcb[pcbidx].current_pid, SIGSTOP);
-
             int t;
             pausetime = getClk();
             pid_t result = waitpid(RR_pcb[pcbidx].current_pid,&t,WUNTRACED);
@@ -202,50 +256,68 @@ void RR(int process_count, int quantum) {
                 RR_pcb[pcbidx].WTA = RR_pcb[pcbidx].TA / RR_runningProcess.runtime;
                 RR_completed++;
                 RR_pcb[pcbidx].remainingtime = 0;
-                Run = false;    
-                addToLog(RR_pcb[pcbidx],getClk());        
-               
+                Run = false; 
+                freeMem(mem,RR_runningProcess.id);
+                addToLog(RR_pcb[pcbidx],getClk());
+
+                MemBlock* allocate = NULL;
+                while (!isCircularQueueEmpty(BlockList))
+                {
+                    allocate = allocateMem(mem,BlockList->array[BlockList->front].memsize,BlockList->array[BlockList->front].id);
+                    if (allocate != NULL)
+                    {
+                        Process p = dequeueCircularQueue(BlockList);
+                        enqueueCircularQueue(RR_queue,p);
+                        RR_pcb[p.id - 1].memblock = allocate;
+                    }
+                    else
+                        break; 
+                } 
             }
         }
 
-        if (!isCircularQueueEmpty(RR_queue)) {
+        
 
-            if(!Run) {
-                util += getClk() - pausetime;          
-                RR_runningProcess = dequeueCircularQueue(RR_queue);
-                pcbidx = RR_runningProcess.id - 1;
+        if (!isCircularQueueEmpty(RR_queue) && !Run) {
+            util += getClk() - pausetime;
+            RR_runningProcess = dequeueCircularQueue(RR_queue);
+            pcbidx = RR_runningProcess.id - 1;
+            current_time = getClk();
+
+            if(RR_pcb[pcbidx].remainingtime > quantum)
+                RR_startTime = current_time;
+            else
+                RR_startTime = current_time - (quantum - RR_pcb[pcbidx].remainingtime);
+
+                
+            if (RR_pcb[pcbidx].processstate == stopped)
+            { 
                 Run = true;
-                current_time = getClk();
-                if(RR_pcb[pcbidx].remainingtime > quantum)
-                    RR_startTime = current_time;
-                else
-                    RR_startTime = current_time - (quantum - RR_pcb[pcbidx].remainingtime);
+                RR_pcb[pcbidx].processstate = resumed;
+                kill(RR_pcb[pcbidx].current_pid,SIGCONT);
+                addToLog(RR_pcb[pcbidx],getClk());      
+            }
+            else
+            {
+                RR_pcb[pcbidx].current_pid = fork();
 
-                if(RR_pcb[pcbidx].processstate == stopped)
-                {
-                    RR_pcb[pcbidx].processstate = resumed;
-                    kill(RR_pcb[pcbidx].current_pid,SIGCONT);
-                    addToLog(RR_pcb[pcbidx],getClk());      
+                if (RR_pcb[pcbidx].current_pid == 0) {
+                    char rt[10], id[10];
+                    sprintf(rt, "%d", RR_pcb[pcbidx].remainingtime);
+                    sprintf(id, "%d", RR_runningProcess.id);
+                    execl("./process.out", "./process.out", rt, id, NULL);
                 }
-                else
+                else 
                 {
-                    RR_pcb[pcbidx].current_pid = fork();
-                    if (RR_pcb[pcbidx].current_pid == 0) {
-                        char rt[10], id[10];
-                        sprintf(rt, "%d", RR_pcb[pcbidx].remainingtime);
-                        sprintf(id, "%d", RR_runningProcess.id);
-                        execl("./process.out", "./process.out", rt, id, NULL);
-                    }
-                    else {
-                        RR_pcb[pcbidx].waitingtime = current_time - RR_runningProcess.arrival_time;
-                        RR_pcb[pcbidx].processstate = running;
-                    
-                        addToLog(RR_pcb[pcbidx],getClk());
-                    }
-                }
+                    RR_pcb[pcbidx].waitingtime = current_time - RR_runningProcess.arrival_time;
+                    RR_pcb[pcbidx].processstate = running;
+                    Run = true;
+                    addToLog(RR_pcb[pcbidx],getClk());
+                }    
             }
         }
     }
+
 }
 
 //////////////////////////////////// Shortest Remaining Time Next //////////////////////////////////////
@@ -261,6 +333,71 @@ void SRTN(int process_count)
         6- if the next process was never forked before, fork it, else resume it
     */
 }
+
+
+//----------------------allocate memory block----------------------
+MemBlock* allocateMem(MemBlock* root, int size, int process_id)
+{ 
+    if(!root || !(root->isFree) || root->bytes < size || size < 2)
+        return NULL;
+    
+    int nearestPower = (int)(ceil(log2(size)));
+
+    //if size is equal to block no need to split
+    if(root->power == nearestPower && root->isFree && !root->isSplit)
+    {
+        root->isFree = false;
+        root->process_id = process_id;
+        MemLog(*root,getClk());
+        return root;
+    }
+
+    // if not and the block was never split before
+    if(!(root->isSplit) && root->isFree)
+    {
+        if(!root->left)
+            root->left = createMemBlock(root->start, root->bytes/2, root->power - 1);
+        if(!root->right)
+            root->right = createMemBlock((root->left->end) + 1, root->bytes/2, root->power - 1);
+
+        root->isSplit = true;
+    }
+
+    // where to allocate check left first is free allocate no see right
+    MemBlock* found = allocateMem(root->left, size, process_id);
+    if(!found)
+        found = allocateMem(root->right, size, process_id);
+
+    //when found recursively set isSplit to true
+    if(found)
+        root->isSplit = true;
+
+    return found;
+}
+
+//----------------------Free memory block----------------------
+bool freeMem(MemBlock* root, int process_id) {
+    if (root == NULL) 
+        return false;
+
+    //if found
+    if (root->process_id == process_id) {
+        root->isFree = true;
+        root->process_id = -1;
+        MemLog(*root,getClk());
+        return true;
+    }
+    //Traverse the tree in case the block is not found
+    bool leftFreed = freeMem(root->left, process_id);
+    bool rightFreed = freeMem(root->right, process_id);
+
+    //recursively merge if both children are free
+    if (leftFreed || rightFreed)
+        mergeMem(root);
+
+    return (leftFreed || rightFreed);
+}
+
 void addToLog(struct PCB pcblock, int time)
 {
     int process_id = pcblock.p.id;
@@ -297,6 +434,23 @@ void addToLog(struct PCB pcblock, int time)
     
     fprintf(logFile,"\n");
 }
+void MemLog(MemBlock mem, int time)
+{
+    //#At time x allocated y bytes for process z from i to j
+     int size = mem.bytes;
+     int start = mem.start;
+     int end = mem.end;
+     int process_id = mem.process_id;
+   
+     if (mem.isFree)
+     {
+         fprintf(MemFile, "At time %d freed %d bytes from %d to %d\n", time, size, start, end);
+     }
+     else
+     {
+         fprintf(MemFile, "At time %d allocated %d bytes for process %d from %d to %d\n", time, size, process_id, start, end);
+     }
+}
 
 void addToPerf(struct PCB *pcb, int process_count)
 {
@@ -304,8 +458,8 @@ void addToPerf(struct PCB *pcb, int process_count)
     double avgWait = 0, avgWTA = 0, stdWTA = 0,totalexc=0 ;
     int arrive;
     int totalrun;
-    finish=pcb[0].finishedtime;
-    arrive=pcb[0].p.arrival_time;
+    finish = pcb[0].finishedtime;
+    arrive = pcb[0].p.arrival_time;
     for(int i = 0; i < process_count; i++)
     {
         
@@ -334,10 +488,17 @@ void addToPerf(struct PCB *pcb, int process_count)
 void closeScheduler(int sig_num)
 {
     printf("Cleaning up resources as Scheduler\n");
+
+
+    if(mem)
+        free(mem);
+
     if(logFile)
         fclose(logFile);
     if(perfFile)
         fclose(perfFile);
+    if(MemFile)
+        fclose(MemFile);
 
     if(pcb)
         free(pcb);
@@ -350,7 +511,11 @@ void closeScheduler(int sig_num)
     destroyClk(false);
     
     destroyCircularQueue(RR_queue);
+    destroyPriorityQueue(BlocklistPri);
+    destroyCircularQueue(BlockList);
 
     if(RR_pcb)
          free(RR_pcb);
+
+    exit(0);
 }
